requirements-flask.txt
# Core Dependencies
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

# AI Integration
openai==1.54.3

# Data & CSV
pandas==2.3.3
numpy==2.2.6

# Lob API
lob==4.5.4

# PDF Generation
reportlab==4.4.4
python-docx==1.2.0
PyPDF2==3.0.1

# Environment
python-dotenv==1.1.1

# Date handling
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# HTTP & API
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11

# XML/Data parsing
lxml==6.0.2
xmltodict==1.0.2

# Image handling
pillow==12.0.0

# Utilities
six==1.17.0
typing_extensions==4.15.0
